[game] change interface to new style (still in Todo)
