+++ xbmc/games/addons/GameClientInGameSaves.cpp
@@ -37,7 +37,7 @@ using namespace GAME;
 #define INGAME_SAVES_EXTENSION_SAVE_RAM ".sav"
 #define INGAME_SAVES_EXTENSION_RTC      ".rtc"
 
-CGameClientInGameSaves::CGameClientInGameSaves(CGameClient* addon, const KodiToAddonFuncTable_Game* dllStruct) :
+CGameClientInGameSaves::CGameClientInGameSaves(CGameClient* addon, const AddonInstance_Game* dllStruct) :
   m_gameClient(addon),
   m_dllStruct(dllStruct)
 {
@@ -85,7 +85,7 @@ void CGameClientInGameSaves::Load(GAME_MEMORY memoryType)
 
   try
   {
-    m_dllStruct->GetMemory(memoryType, &gameMemory, &size);
+    m_dllStruct->toAddon.get_memory(m_dllStruct, memoryType, &gameMemory, &size);
   }
   catch (...)
   {
@@ -126,7 +126,7 @@ void CGameClientInGameSaves::Save(GAME_MEMORY memoryType)
 
   try
   {
-    m_dllStruct->GetMemory(memoryType, &gameMemory, &size);
+    m_dllStruct->toAddon.get_memory(m_dllStruct, memoryType, &gameMemory, &size);
   }
   catch (...)
   {
