+++ xbmc/games/addons/GameClientKeyboard.h
@@ -21,7 +21,7 @@
 
 #include "input/keyboard/IKeyboardHandler.h"
 
-struct KodiToAddonFuncTable_Game;
+struct AddonInstance_Game;
 
 namespace KODI
 {
@@ -49,7 +49,7 @@ namespace GAME
      * \param dllStruct The emulator or game to which the events are sent.
      * \param inputProvider The interface providing us with keyboard input.
      */
-    CGameClientKeyboard(const CGameClient* gameClient, const KodiToAddonFuncTable_Game* dllStruct, KEYBOARD::IKeyboardInputProvider *inputProvider);
+    CGameClientKeyboard(const CGameClient* gameClient, const AddonInstance_Game* dllStruct, KEYBOARD::IKeyboardInputProvider *inputProvider);
 
     /*!
      * \brief Destructor unregisters from keyboard events from CInputManager.
@@ -63,7 +63,7 @@ namespace GAME
   private:
     // Construction parameters
     const CGameClient* const m_gameClient;
-    const KodiToAddonFuncTable_Game* const m_dllStruct;
+    const AddonInstance_Game* const m_dllStruct;
     KEYBOARD::IKeyboardInputProvider *const m_inputProvider;
   };
 }
