+++ xbmc/games/addons/GameClientMouse.cpp
@@ -21,7 +21,7 @@
 #include "GameClientMouse.h"
 #include "GameClient.h"
 #include "GameClientTranslator.h"
-#include "addons/kodi-addon-dev-kit/include/kodi/kodi_game_types.h"
+#include "addons/kodi-addon-dev-kit/include/kodi/addon-instance/Game.h"
 #include "input/mouse/IMouseInputProvider.h"
 #include "input/Key.h"
 #include "utils/log.h"
@@ -29,7 +29,7 @@
 using namespace KODI;
 using namespace GAME;
 
-CGameClientMouse::CGameClientMouse(const CGameClient* gameClient, const KodiToAddonFuncTable_Game* dllStruct, MOUSE::IMouseInputProvider *inputProvider) :
+CGameClientMouse::CGameClientMouse(const CGameClient* gameClient, const AddonInstance_Game* dllStruct, MOUSE::IMouseInputProvider *inputProvider) :
   m_gameClient(gameClient),
   m_dllStruct(dllStruct),
   m_inputProvider(inputProvider),
@@ -68,7 +68,7 @@ bool CGameClientMouse::OnMotion(const std::string& relpointer, int dx, int dy)
 
   try
   {
-    bHandled = m_dllStruct->InputEvent(&event);
+    bHandled = m_dllStruct->toAddon.input_event(m_dllStruct, &event);
   }
   catch (...)
   {
@@ -98,7 +98,7 @@ bool CGameClientMouse::OnButtonPress(const std::string& button)
 
   try
   {
-    bHandled = m_dllStruct->InputEvent(&event);
+    bHandled = m_dllStruct->toAddon.input_event(m_dllStruct, &event);
   }
   catch (...)
   {
@@ -120,7 +120,7 @@ void CGameClientMouse::OnButtonRelease(const std::string& button)
 
   try
   {
-    m_dllStruct->InputEvent(&event);
+    m_dllStruct->toAddon.input_event(m_dllStruct, &event);
   }
   catch (...)
   {
