+++ xbmc/games/addons/GameClientMouse.h
@@ -21,7 +21,7 @@
 
 #include "input/mouse/IMouseInputHandler.h"
 
-struct KodiToAddonFuncTable_Game;
+struct AddonInstance_Game;
 
 namespace KODI
 {
@@ -49,7 +49,7 @@ namespace GAME
      * \param dllStruct The emulator or game to which the events are sent.
      * \param inputProvider The interface providing us with mouse input.
      */
-    CGameClientMouse(const CGameClient* gameClient, const KodiToAddonFuncTable_Game* dllStruct, MOUSE::IMouseInputProvider *inputProvider);
+    CGameClientMouse(const CGameClient* gameClient, const AddonInstance_Game* dllStruct, MOUSE::IMouseInputProvider *inputProvider);
 
     /*!
      * \brief Destructor unregisters from mouse events from CInputManager.
@@ -65,7 +65,7 @@ namespace GAME
   private:
     // Construction parameters
     const CGameClient* const m_gameClient;
-    const KodiToAddonFuncTable_Game* const m_dllStruct;
+    const AddonInstance_Game* const m_dllStruct;
     MOUSE::IMouseInputProvider *const m_inputProvider;
     const std::string m_controllerId;
   };
